Made GUI, easier to grasp, added right-click popup and menu bar.
Changed the way that the software works.
Now when you log in, you can see all your files, double clicking on a file
now opens in your default editor, after closing it, the modified file is
uploaded immediately to the server.
Server currently doesn't check if you have the right to modify the file.
I will add this feature soon.

Signed-off-by: Robert Baronescu <baronescu.robert@gmail.com>